Refactor: Post & MyPost 병합 후 조건부 렌더링으로 변경
