--- src/pages/MyPost/styles.tsx
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-
-export const ModalContainer = styled.div`
-	background: white;
-	border-radius: 10px 10px 0 0;
-	max-width: 512px;
-	position: flex;
-	height: 377px;
-	flex-shrink: 0;
-`;
