+++ src/pages/Post/PostBase/LikeCommentBottomSheetContent/CommentItem/dto.ts
@@ -1,4 +1,4 @@
-import { Comment } from '../../../../apis/post-comment/dto';
+import { Comment } from '../../../../../apis/post-comment/dto';
 
 export interface CommentItemProps {
 	comment: Comment;
