+++ src/pages/Post/PostBase/LikeCommentBottomSheetContent/MenuButtonList/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { MenuButtonListProps } from './dto';
 import { MenuListWrapper, MenuListContainer, MenuButtonItem } from './styles';
-import { StyledText } from '../../../Text/StyledText';
+import { StyledText } from '../../../../../components/Text/StyledText';
 
 const MenuButtonList: React.FC<MenuButtonListProps> = ({ items, onClose, position }) => {
 	const handleWrapperClick = () => {
