+++ src/pages/Post/PostBase/dto.ts
